+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx.hs
@@ -40,7 +40,6 @@ module Cardano.Ledger.Api.Tx (
   bodyTxL,
   witsTxL,
   auxDataTxL,
-  validateScript,
   sizeTxF,
   getMinFeeTx,
   setMinFeeTx,